Trying to make a daily calendar view
