+++ app/src/main/AndroidManifest.xml
@@ -14,19 +14,6 @@
         android:theme="@style/Theme.Applikation"
         tools:targetApi="31">
         <activity
-            android:name=".Calendar"
-            android:exported="false"/>
-        <activity
-            android:name=".CurrentLog"
-            android:exported="false" />
-        <activity
-            android:name=".Chat"
-            android:exported="false" />
-        <activity
-            android:name=".Log"
-            android:windowSoftInputMode="adjustPan"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
@@ -35,6 +22,19 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".calendar.Calendar"
+            android:exported="false" />
+        <activity
+            android:name=".log.CurrentLog"
+            android:exported="false" />
+        <activity
+            android:name=".chat.Chat"
+            android:exported="false" />
+        <activity
+            android:name=".log.Log"
+            android:exported="false"
+            android:windowSoftInputMode="adjustPan" />
     </application>
 
 </manifest>
