+++ app/src/main/java/com/miun/applikation/chat/ChatAdapter.java
@@ -1,4 +1,4 @@
-package com.miun.applikation;
+package com.miun.applikation.chat;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -7,6 +7,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.miun.applikation.R;
 
 import java.util.List;
