+++ app/src/main/java/com/miun/applikation/log/LogAdapter.java
@@ -1,16 +1,15 @@
-package com.miun.applikation;
+package com.miun.applikation.log;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import java.text.BreakIterator;
+import com.miun.applikation.R;
+
 import java.util.List;
 
 public class LogAdapter extends RecyclerView.Adapter<LogAdapter.LogViewHolder> {
