+++ app/src/main/java/com/miun/applikation/utils/HelperFunctions.java
@@ -1,4 +1,4 @@
-package com.miun.applikation;
+package com.miun.applikation.utils;
 
 import android.app.Activity;
 import android.view.inputmethod.InputMethodManager;
