+++ app/src/main/res/layout/calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Calendar">
+    tools:context=".calendar.Calendar">
 
     <Button
         android:id="@+id/goBackCalendar"
