+++ app/src/main/res/layout/log.xml
@@ -5,7 +5,7 @@
     android:id="@+id/linearLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Chat">
+    tools:context=".chat.Chat">
 
     <Button
         android:id="@+id/goBackLog"
